Goals and problems
I'm trying to create a project that uses a motion sensor and ultrasonic distance sensor as inputs to arm a "person detected alarm". The ultrasonic sensor is mainly to prevent false triggers, but I like being able to control a distance that the system will activate at as well. The general idea is as follows:
-System turns on, LED turns solid after a "warm-up" phase (30 seconds).
-Once motion is detected, wait for the distance to either change from a set value (such as a tripwire), or be within a certain distance. 
- IF motion is active AND distance is triggered, 2 relays are activated, the LED blinks on and off, and nothing else will happen until I manually reset the system via a button. 

Issues i'm having: 
- Once I turn it on, I'm unsure if it is fully warmed up and waiting. Solution: I want to add an OLED and menu with either a keypad or rotary encoder to toggle through and select actions such as "Arm System", "Set Distance", and "Disarm System".
- Relay 1 reactivates every 25 seconds, regardless of if I hit reset or not. Needs to wait until I reset. Not sure why it keeps activating. 

The code I have in the first iteration is very unorganized, and the timings are almost irrelevant. Initially, I wanted the relays to be a few MS apart, last 6 seconds, and rearm again in 1 minute. Now, I just want the duration of the relays to be 6 seconds and that's all. 

Once that's cleaned up, I just want to include the menu. 

Any help is appreciated.